fix: Исправлено переполнение типа telegram_id (int32 → bigint)
Проблема:
- Telegram ID может превышать максимум int32 (2,147,483,647)
- Реальные примеры: 6,401,854,279 и 5,666,608,489
- Ошибка: "invalid input for query argument: value out of int32 range"
- Новые пользователи не могли использовать бота

Исправления в коде:
1. Изменён тип telegram_id в модели User с Integer на BigInteger
2. Исправлены вызовы сервисов в meal_plan.py:
   - Передавался user.telegram_id вместо user.id
   - user.id - это ID записи в БД (INTEGER)
   - user.telegram_id - это реальный Telegram ID (теперь BIGINT)

Файлы миграции БД:
- migrations/migrate_telegram_id_to_bigint.sql - SQL скрипт
- migrations/apply_telegram_id_bigint_migration.py - Python скрипт
- migrations/TELEGRAM_ID_MIGRATION_README.md - документация

Затронуты:
- app/models/user.py - изменён тип поля telegram_id
- app/bot/handlers/meal_plan.py - исправлено 5 вызовов сервисов

+++ app/models/user.py
@@ -1,7 +1,7 @@
 """
 Модель пользователя
 """
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, Enum as SQLEnum
+from sqlalchemy import Column, Integer, BigInteger, String, Float, Boolean, DateTime, Enum as SQLEnum
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -61,7 +61,7 @@ class User(Base):
 
     # ID
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     language_code = Column(String(10), default="ru")
 
     # Зашифрованные персональные данные (152-ФЗ)
